requirements.txt
fastapi
uvicorn
transformers
unsloth
accelerate
bitsandbytes
torch
llama-cpp-python
pydantic
